Stage (#121)

+++ src/cver/client/__init__.py
@@ -10,10 +10,9 @@ import tempfile
 
 import requests
 
-from cver.shared.utils import misc as s_misc
 from cver.shared.utils import xlate
 from cver.api.version import version as __version__
-from cver.client.utils.read_config import ReadConfig
+from cver.client.utils.config import Config
 # from cver.shared.utils.log_config import log_config
 
 
@@ -28,37 +27,18 @@ class Client:
         """Initialize the CverClient with the client_id, api_key and/or api_url. If not supplied
         environmental vars will be attempted.
         """
-        if client_id:
-            self.client_id = client_id
-        else:
-            self.client_id = os.environ.get("CVER_CLIENT_ID")
+        self.config = Config().get(client_id, api_key, api_url)
+        self.api_url = self.config["api_url"]
+        self.client_id = self.config["client_id"]
+        self.api_key = self.config["api_key"]
+        self.api_host_name = self.config["api_host_name"]
 
-        if api_key:
-            self.api_key = api_key
-        else:
-            self.api_key = os.environ.get("CVER_API_KEY")
-
-        if api_url:
-            self.api_url = api_url
-        else:
-            self.api_url = os.environ.get("CVER_API_URL")
-        self.api_url = s_misc.strip_trailing_slash(self.api_url)
-        self.user_agent = "CverClient/%s" % __version__
-        self.config = ReadConfig().read()
         self.headers = {}
-        self.api_host = os.environ.get("CVER_API_HOST")
-        self.api_url = s_misc.strip_trailing_slash(self.api_url)
-
-        cver_api_host = os.environ.get("CVER_API_HOST")
-        if cver_api_host:
-            self.api_host = cver_api_host
-            self.headers["Host"] = os.environ.get("CVER_API_URL")
-            self.api_url = s_misc.strip_trailing_slash(cver_api_host)
-            self.headers["Host"] = self.headers["Host"].replace("http://", "")
-            self.headers["Host"] = self.headers["Host"].replace("https://", "")
+        if self.api_host_name:
+            self.headers["Host"] = self.api_host_name
+        self.headers["User-Agent"] = "CverClient/%s" % __version__
 
         self.token = ""
-        self.token_path = ""
         self.response = None
         temp_dir = tempfile.gettempdir()
         self.token_file = os.path.join(temp_dir, "cver-token")
@@ -68,8 +48,12 @@ class Client:
 
     def login(self, skip_local_token: bool = False) -> bool:
         """Login to the Cver API."""
+        if self.login_attempts >= self.max_login_attempts:
+            logging.crticial("Reached max loging attempts (%s), quitting")
+            return False
         logging.debug("Logging into Cver Api: %s" % self.api_url)
         if not skip_local_token and self._open_valid_token():
+            # logging.info("Using existing token: %s" % self.token_file)
             return True
         if not self._determine_if_login():
             return False
@@ -83,7 +67,14 @@ class Client:
             "url": f"{self.api_url}/auth",
         }
         request_args["headers"].update(self.headers)
+        logging.info("Making auth request")
         response = requests.request(**request_args)
+        if response.status_code == 401:
+            logging.warning("Received unauthorized status code 401 logging in")
+            self.destroy_token()
+            self.login_attempts += 1
+            return False
+
         if response.status_code == 403:
             logging.error("Received status code %s logging in" % response.status_code)
             self.login_attempts += 1
@@ -103,23 +94,11 @@ class Client:
         the response json back.
         @todo: Break this apart - too complex
         """
-        self.login()
-        # if not self.token:
-        #     self.login()
-        headers = {
-            "token": self.token,
-            "content-type": "application/json",
-            "User-Agent": self.user_agent
-        }
-        if self.api_host:
-            headers["Host"] = self.api_host
-        request_args = {
-            "headers": headers,
-            "method": method,
-            "url": f"{self.api_url}/{url}"
-        }
+        if not self.token:
+            logging.debug("Attempt login")
+            self.login()
 
-        request_args["headers"].update(self.headers)
+        request_args = self._get_base_request_args(url, method)
 
         if request_args:
             if method == "GET":
@@ -180,8 +159,15 @@ class Client:
         response = self.make_request("submit-scan", "POST", payload)
         print(response)
 
-    def destroy_token(self):
-        """Delete a local token from temp space."""
+    def destroy_token(self) -> bool:
+        """Delete a local token from temp space
+        :unit-test: TestClient
+        """
+        logging.debug("Delete Cver Token: %s" % self.token_file)
+        self.token = None
+        if not os.path.exists(self.token_file):
+            logging.error("Cant delete token that doesnt exist")
+            return True
         os.remove(self.token_file)
         logging.info("Deleted local Cver token.")
         return True
@@ -199,8 +185,10 @@ class Client:
         return True
 
     def _save_token(self):
-        """Save a token to a local tempfile location."""
-        logging.info(f"Temp Dir is: {self.token_file}")
+        """Save a token to a local tempfile location.
+        :unit-test: TestClientInit::test___save_token
+        """
+        # logging.info(f"Temp Dir is: {self.token_file}")
         if not self.token:
             logging.error("No token to save.")
             return False
@@ -223,6 +211,21 @@ class Client:
         self.token = token_data
         return True
 
+    def _get_base_request_args(self, url: str, method: str) -> dict:
+        """Get the base request args for requests on the Cver Api.
+        :unit-test: TestClient::test___get_base_request_args
+        """
+        request_args = {
+            "headers": {
+                "token": self.token,
+                "content-type": "application/json",
+            },
+            "method": method,
+            "url": f"{self.api_url}/{url}"
+        }
+        request_args["headers"].update(self.headers)
+        return request_args
+
     def _handle_error(self, response, request_args) -> bool:
         url = request_args["url"]
         if "token" in request_args:
@@ -236,4 +239,4 @@ class Client:
         return False
 
 
-# End File: cver/src/shared/cver_client/__init__.py
+# End File: cver/src/shared/client/__init__.py
